feat: integrate orb did method
Adds `orb`, a new service, which runs a small orb network and a
universal resolver orb driver instance.

Integrates the service with the afgo backchannel, by having the service
create an orb did and save the value in a data folder for the afgo backchannel
to load.

Remember to start orb with `./manage service start orb` for tests where afgo
needs a public did.

Adds a general way to load files into agent containers:
A directory named <agent-name>.data (eg, afg-master.data), if present within the
backchannel's folder, will be mounted as a volume within the given agent,
to the path `/data-mount`.

Signed-off-by: Filip Burlacu <filip.burlacu@securekey.com>